requirements.txt
absl-py==0.11.0
astunparse==1.6.3
attrs==19.3.0
cached-property==1.5.2
cachetools==3.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dm-tree==0.1.5
fastcluster==1.2.4
flatbuffers==1.12
gast==0.4.0
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.34.1
h5py==3.1.0
html==1.10
html-parser==0.2
HTMLParser==0.0.2
idna==2.10
importlib-metadata==3.0.0
jax==0.2.13
jaxlib==0.1.67
joblib==1.0.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
mpmath==1.1.0
networkx==2.6.3
nltk==3.5
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
Pillow==8.0.1
ply==3.11
portpicker==1.3.9
protobuf==3.14.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scipy==1.7.1
semantic-version==2.8.5
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.5.0
tensorflow-addons==0.15.0
tensorflow-estimator==2.5.0
tensorflow-federated==0.19.0
tensorflow-model-optimization==0.5.0
tensorflow-privacy==0.5.2
termcolor==1.1.0
tqdm==4.62.0
typeguard==2.13.3
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
